api/requirements.txt
pydantic==1.9.0
uvicorn==0.17.6
fastapi==0.75.1
sqlalchemy==1.4.36
psycopg2==2.9.3
bcrypt==3.2.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
firebase-admin==5.2.0
Unidecode==1.3.4